Move PublicSearchResults into public directory
The PublicSearchResultsComponent was in its own directory outside of the
directory with the rest of the PublicModule components. Move it into the
proper components directory and update imports to point to the new
location.
