+++ src/app/public/components/public-search-results/public-search-results.component.scss
