+++ src/app/public/components/public-search-results/public-search-results.component.ts
@@ -1,5 +1,5 @@
 /* @format */
-import { SearchService } from '../search/services/search.service';
+import { SearchService } from '../../../search/services/search.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
